add limited framework for specific orion/homeworld locations
add setSpecific and indexWorld in GalaxyShape to allow map generation to set specific locations for orion and empire homeworlds.

experimental proof-of-concept in EllipticalShape, currently only for Circular with Void 4 and 100 or more Stars.

empire homeworlds should be distributed at equal positions along the ring, with orion in the middle (for large maps, a current bug does not ensure orion to be in the center).